refactor: share token map code
