+++ src/lib/hooks/useTokenList/utils.ts
@@ -1,0 +1,32 @@
+import { TokenInfo, TokenList } from '@uniswap/token-lists'
+import { WrappedTokenInfo } from 'state/lists/wrappedTokenInfo'
+
+export type TokenMap = Readonly<{ [tokenAddress: string]: { token: WrappedTokenInfo; list?: TokenList } }>
+export type ChainTokenMap = Readonly<{ [chainId: number]: TokenMap }>
+
+type Mutable<T> = {
+  -readonly [P in keyof T]: Mutable<T[P]>
+}
+
+const mapCache = typeof WeakMap !== 'undefined' ? new WeakMap<TokenList | TokenInfo[], ChainTokenMap>() : null
+
+export function tokensToChainTokenMap(tokens: TokenList | TokenInfo[]): ChainTokenMap {
+  const cached = mapCache?.get(tokens)
+  if (cached) return cached
+
+  const [list, infos] = Array.isArray(tokens) ? [undefined, tokens] : [tokens, tokens.tokens]
+  const map = infos.reduce<Mutable<ChainTokenMap>>((map, info) => {
+    const token = new WrappedTokenInfo(info, list)
+    if (map[token.chainId]?.[token.address] !== undefined) {
+      console.warn(`Duplicate token skipped: ${token.address}`)
+      return map
+    }
+    if (!map[token.chainId]) {
+      map[token.chainId] = {}
+    }
+    map[token.chainId][token.address] = { token, list }
+    return map
+  }, {}) as ChainTokenMap
+  mapCache?.set(tokens, map)
+  return map
+}
